fead: Add invoke_messagemf_borrowed2
This is just a copy paste of invoke_messagemf_borrowed and then renaming
it invoke_messagemf_borrowed2. I also reran `./gen_asm.sh` and the
previously generated code asm/*.txt file are identical and we only see
the new asm/infoke_messagemf_borrowed2.txt file.

Alas, there is non-trivial differences in the output of
`taskset 1 cargo bench`. Of note the number of instructions
were identical, but in most of the other lines there are differences :(